Refacto: rename github.py module into github_utils.py to avoid Import error with circular import
