+++ bera/historical_extract_all_beras.py
@@ -5,7 +5,7 @@ import time
 
 from utils.bulletin import Bulletin
 from utils.common import init_logger, MASSIFS
-from utils.github import init_repo, commit_many_files_and_push, \
+from utils.github_utils import init_repo, commit_many_files_and_push, \
     update_and_add_file_to_commit
 
 logger = init_logger(logging.DEBUG)
